# Avoid usage of __builtin_popcountll when -simdutf is unset