--- a/.github/workflows/simdutf-flag-windows.yml
+++ b/.github/workflows/simdutf-flag-windows.yml
@@ -1,0 +1,40 @@
+name: no-simdutf-windows-static
+on:
+  - push
+  - pull_request
+
+defaults:
+  run:
+    shell: bash
+
+jobs:
+  build:
+    runs-on: ${{ matrix.os }}
+    strategy:
+      matrix:
+        os: ['windows-latest']
+        ghc: ['9.2.3']
+      fail-fast: false
+    steps:
+      - uses: actions/checkout@v2
+      - uses: haskell/actions/setup@v1
+        id: setup-haskell-cabal
+        with:
+          ghc-version: ${{ matrix.ghc }}
+      - name: Update cabal package database
+        run: cabal update
+      - uses: actions/cache@v2
+        name: Cache cabal stuff
+        with:
+          path: ${{ steps.setup-haskell-cabal.outputs.cabal-store }}
+          key: ${{ runner.os }}-${{ matrix.ghc }}
+
+      - name: Build Tests
+        run: |
+          cabal build test:tests --enable-tests --enable-benchmarks -f-simdutf;
+      - name: Run Tests
+        run: |
+          $bin = cabal list-bin test:tests
+          $env:PATH = ''
+          & "$bin"
+        shell: pwsh