--- a/text.cabal
+++ b/text.cabal
@@ -193,10 +193,6 @@ library
     if impl(ghc >= 9.2.2)
       ghc-options: -fcheck-prim-bounds
 
-  -- https://gitlab.haskell.org/ghc/ghc/-/issues/19900
-  if os(windows) && impl(ghc < 9.3)
-    extra-libraries: gcc_s
-
   default-language: Haskell2010
   default-extensions:
     NondecreasingIndentation