Also avoid linking against gcc/gcc_s on all platforms.

This works around https://gitlab.haskell.org/ghc/ghc/-/issues/21787
and https://gitlab.haskell.org/ghc/ghc/-/issues/19900 which cause
problems when GHC's RTS linker tries to load `text`, which occurs if
you use a statically linked GHC to compile a file with a TH splice that
depends on `text`.

Since we don't require SSE4.2 to build `text -simdutf`, this shouldn't
be much of a pessimisation.

Fixes https://github.com/haskell/text/issues/450

This is meant to be a temporary workaround for https://gitlab.haskell.org/ghc/ghc/-/issues/21787 while we work on a robust method for properly exposing all GCC symbols from the RTS linker.

However, older versions of GHC (particularly the 9.0 series and earlier) which won't be patched still need a workaround so that `text-2.0` is usable under all configurations.

It is also problematic to include `extra-libraries: gcc_s` when compiling with clang.