@wz1000 2022-07-11
@Bodigrim I've modified the `-simdutf` CI job to test on windows (which is the only GHC distribution we ship with a statically linked GHC). I also modified it to test GHC-9.2.3 instead of `latest` because currently `latest` resolves to `9.2.2`, and chocolatey includes a workaround on that version which also masks the text issue.

See https://github.com/haskell/text/runs/7279337599 (which is a run from #454) for an example of the failure without this patch.