@wz1000 2022-07-12
Unfortunately the static alpine GHC binaries are not usable because of https://gitlab.haskell.org/ghc/ghc/-/issues/21844

I'm stumped by this, not sure how to add a non windows CI job that demonstrates the problem.