@wz1000 2022-07-12
I guess I can try to run it in an alpine container and it should work. I'll try this tomorrow.