@Bodigrim 2022-07-12
> I'm stumped by this, not sure how to add a non windows CI job that demonstrates the problem.

As I suggested in https://github.com/haskell/text/issues/450#issuecomment-1180871134, try to reproduce `bytestring` CI with respect to running Windows job with a clean PATH: https://github.com/haskell/bytestring/blob/22b36125ac52605e807b7b96ef31e8f087248f17/.github/workflows/ci.yml#L93-L99
If this proves that `extra-libraries: gcc_s` is not a good solution, we are all set, no need for Alpine.