@wz1000 2022-07-14
I've fixed the alpine job and also attempted to add a windows job to perform the check you added in bytestring, but that job seems to fail in a different way to the intended failure mode. See https://github.com/haskell/text/runs/7337239414?check_suite_focus=true