@Bodigrim 2022-07-14
The simplest setup I can come up is https://github.com/Bodigrim/text/tree/purge-path. After the first commit the build succeeds, but as soon as we purge PATH, if fails (the error message is just error code 1, that's fine). If this setup is enough, I'd rather avoid Alpine job with unreleased GHC version. 

Now I don't quite understand what does it have to do with `simdutf` flag. The code is used unconditionally for any `length` / `drop` / `take`, so its performance is absolutely crucial. Please acompany your patch with benchmark results (`cabal bench --benchmark-options='-p length'`).