@bgamari 2022-07-15
The failure in #454 is due not to `text` but rather `bytestring`, which `text` links against. See https://github.com/haskell/bytestring/issues/497.

This should be fixed in the `bytestring` shipped with 9.2.3.