@bgamari 2022-07-15
#454 also neglects to disable the `simdutf8` flag, which introduces a dependency on `libstdc++`, which will of course fail for the same reason described in the last paragraph of [GHC #20878](https://gitlab.haskell.org/ghc/ghc/-/issues/20878).