@Bodigrim 2022-07-17
@wz1000 could you please check performance impact of your change?