@wz1000 2022-07-18
Here are the results of running `cabal bench --benchmark-options='-p length'` with GHC 9.2.3

Before:

```
All
  Pure
    tiny
      length
        cons
          Text:     OK (0.22s)
            22.4 ns ± 1.7 ns
          LazyText: OK (0.82s)
            23.6 ns ± 1.4 ns
        decode
          Text:     OK (0.20s)
            41.2 ns ± 3.6 ns
          LazyText: OK (0.29s)
            124  ns ± 6.2 ns
        drop
          Text:     OK (0.17s)
            33.6 ns ± 3.1 ns
          LazyText: OK (0.18s)
            35.7 ns ± 3.3 ns
        filter
          Text:     OK (0.19s)
            18.5 ns ± 1.7 ns
          LazyText: OK (0.21s)
            40.3 ns ± 3.9 ns
        filter.filter
          Text:     OK (0.23s)
            23.4 ns ± 1.9 ns
          LazyText: OK (0.20s)
            40.3 ns ± 3.0 ns
        init
          Text:     OK (0.31s)
            30.3 ns ± 1.6 ns
          LazyText: OK (0.26s)
            26.6 ns ± 1.8 ns
        intercalate
          Text:     OK (0.29s)
            30.5 ns ± 2.2 ns
          LazyText: OK (0.15s)
            31.5 ns ± 2.8 ns
        intersperse
          Text:     OK (0.27s)
            29.2 ns ± 1.8 ns
          LazyText: OK (0.27s)
            27.9 ns ± 1.7 ns
        map
          Text:     OK (0.88s)
            24.4 ns ± 1.2 ns
          LazyText: OK (0.25s)
            25.2 ns ± 2.3 ns
        map.map
          Text:     OK (0.45s)
            24.1 ns ± 2.0 ns
          LazyText: OK (0.46s)
            24.9 ns ± 700 ps
        replicate char
          Text:     OK (0.40s)
            21.2 ns ± 1.0 ns
          LazyText: OK (0.18s)
            16.8 ns ± 1.4 ns
        replicate string
          Text:     OK (0.25s)
            25.3 ns ± 1.7 ns
          LazyText: OK (0.21s)
            21.3 ns ± 1.8 ns
        take
          Text:     OK (0.45s)
            24.4 ns ± 1.6 ns
          LazyText: OK (0.16s)
            29.1 ns ± 2.7 ns
        tail
          Text:     OK (0.24s)
            23.8 ns ± 1.9 ns
          LazyText: OK (0.25s)
            24.7 ns ± 1.8 ns
        toLower
          Text:     OK (0.27s)
            115  ns ± 8.4 ns
          LazyText: OK (0.24s)
            196  ns ±  13 ns
        toUpper
          Text:     OK (0.14s)
            118  ns ±  12 ns
          LazyText: OK (0.15s)
            228  ns ±  22 ns
        words
          Text:     OK (0.22s)
            21.5 ns ± 1.9 ns
          LazyText: OK (1.29s)
            36.4 ns ± 1.8 ns
        zipWith
          Text:     OK (0.25s)
            25.8 ns ± 2.2 ns
          LazyText: OK (0.25s)
            26.6 ns ± 1.3 ns
    ascii-small
      length
        cons
          Text:     OK (0.78s)
            11.7 μs ± 187 ns
          LazyText: OK (0.22s)
            11.8 μs ± 928 ns
        decode
          Text:     OK (0.67s)
            17.6 μs ± 1.4 μs
          LazyText: OK (0.72s)
            18.7 μs ± 793 ns
        drop
          Text:     OK (0.22s)
            11.9 μs ± 1.2 μs
          LazyText: OK (0.42s)
            11.9 μs ± 502 ns
        filter
          Text:     OK (0.30s)
            139  μs ± 8.3 μs
          LazyText: OK (0.17s)
            145  μs ±  11 μs
        filter.filter
          Text:     OK (0.13s)
            137  μs ±  11 μs
          LazyText: OK (0.17s)
            144  μs ±  11 μs
        init
          Text:     OK (0.20s)
            11.4 μs ± 670 ns
          LazyText: OK (0.22s)
            11.6 μs ± 695 ns
        intercalate
          Text:     OK (0.41s)
            23.2 μs ± 1.5 μs
          LazyText: OK (0.23s)
            25.4 μs ± 2.0 μs
        intersperse
          Text:     OK (0.41s)
            11.7 μs ± 432 ns
          LazyText: OK (0.22s)
            11.6 μs ± 797 ns
        map
          Text:     OK (0.22s)
            11.7 μs ± 698 ns
          LazyText: OK (0.42s)
            11.9 μs ± 363 ns
        map.map
          Text:     OK (0.22s)
            11.8 μs ± 718 ns
          LazyText: OK (0.22s)
            11.9 μs ± 898 ns
        replicate char
          Text:     OK (0.25s)
            23.8 ns ± 1.8 ns
          LazyText: OK (0.21s)
            19.2 ns ± 1.6 ns
        replicate string
          Text:     OK (0.29s)
            28.5 ns ± 1.9 ns
          LazyText: OK (0.25s)
            23.9 ns ± 1.9 ns
        take
          Text:     OK (0.16s)
            7.98 μs ± 669 ns
          LazyText: OK (0.15s)
            7.89 μs ± 699 ns
        tail
          Text:     OK (0.22s)
            11.8 μs ± 947 ns
          LazyText: OK (0.23s)
            11.8 μs ± 735 ns
        toLower
          Text:     OK (0.35s)
            1.30 ms ± 102 μs
          LazyText: OK (0.24s)
            1.74 ms ± 128 μs
        toUpper
          Text:     OK (0.24s)
            1.76 ms ±  91 μs
          LazyText: OK (0.16s)
            2.26 ms ± 203 μs
        words
          Text:     OK (0.18s)
            311  μs ±  27 μs
          LazyText: OK (0.34s)
            622  μs ±  26 μs
        zipWith
          Text:     OK (0.40s)
            11.5 μs ± 797 ns
          LazyText: OK (0.22s)
            11.7 μs ± 918 ns
    ascii
      length
        cons
          Text:     OK (0.56s)
            10.2 ms ± 911 μs
          LazyText: OK (3.44s)
            10.5 ms ± 857 μs
        decode
          Text:     OK (1.80s)
            18.1 ms ± 843 μs
          LazyText: OK (1.13s)
            18.9 ms ± 564 μs
        drop
          Text:     OK (0.95s)
            10.4 ms ± 480 μs
          LazyText: OK (0.46s)
            10.3 ms ± 934 μs
        filter
          Text:     OK (0.57s)
            122  ms ± 4.4 ms
          LazyText: OK (0.49s)
            128  ms ± 5.0 ms
        filter.filter
          Text:     OK (0.57s)
            121  ms ± 3.2 ms
          LazyText: OK (0.59s)
            129  ms ± 4.7 ms
        init
          Text:     OK (0.72s)
            13.6 ms ± 742 μs
          LazyText: OK (0.55s)
            10.2 ms ± 734 μs
        intercalate
          Text:     OK (0.64s)
            21.0 ms ± 1.2 ms
          LazyText: OK (0.36s)
            21.9 ms ± 1.6 ms
        intersperse
          Text:     OK (0.86s)
            11.2 ms ± 1.0 ms
          LazyText: OK (0.74s)
            10.8 ms ± 337 μs
        map
          Text:     OK (0.85s)
            10.5 ms ± 347 μs
          LazyText: OK (0.46s)
            10.3 ms ± 813 μs
        map.map
          Text:     OK (0.59s)
            10.0 ms ± 425 μs
          LazyText: OK (0.71s)
            10.4 ms ± 786 μs
        replicate char
          Text:     OK (2.47s)
            22.8 ns ± 1.8 ns
          LazyText: OK (2.41s)
            18.7 ns ± 1.8 ns
        replicate string
          Text:     OK (2.48s)
            27.3 ns ± 1.9 ns
          LazyText: OK (2.79s)
            23.2 ns ± 1.4 ns
        take
          Text:     OK (0.71s)
            6.69 ms ± 352 μs
          LazyText: OK (1.46s)
            6.93 ms ± 477 μs
        tail
          Text:     OK (0.56s)
            10.2 ms ± 814 μs
          LazyText: OK (0.72s)
            9.94 ms ± 656 μs
        toLower
          Text:     OK (8.66s)
            1.206 s ±  15 ms
          LazyText: OK (4.66s)
            1.485 s ±  25 ms
        toUpper
          Text:     OK (4.55s)
            1.479 s ±  30 ms
          LazyText: OK (5.70s)
            1.832 s ± 5.5 ms
        words
          Text:     OK (0.95s)
            252  ms ± 4.5 ms
          LazyText: OK (1.74s)
            515  ms ± 3.5 ms
        zipWith
          Text:     OK (0.44s)
            9.80 ms ± 773 μs
          LazyText: OK (0.45s)
            9.80 ms ± 907 μs
    english
      length
        cons
          Text:     OK (0.26s)
            643  μs ±  44 μs
          LazyText: OK (0.24s)
            650  μs ±  43 μs
        decode
          Text:     OK (1.12s)
            942  μs ±  23 μs
          LazyText: OK (0.69s)
            1.12 ms ±  66 μs
        drop
          Text:     OK (0.24s)
            655  μs ±  44 μs
          LazyText: OK (0.23s)
            654  μs ±  55 μs
        filter
          Text:     OK (0.28s)
            7.71 ms ± 386 μs
          LazyText: OK (0.15s)
            8.24 ms ± 822 μs
        filter.filter
          Text:     OK (0.15s)
            7.79 ms ± 706 μs
          LazyText: OK (0.30s)
            8.16 ms ± 388 μs
        init
          Text:     OK (0.23s)
            646  μs ±  48 μs
          LazyText: OK (0.24s)
            651  μs ±  46 μs
        intercalate
          Text:     OK (0.22s)
            1.32 ms ± 100 μs
          LazyText: OK (0.24s)
            1.39 ms ±  91 μs
        intersperse
          Text:     OK (0.14s)
            650  μs ±  59 μs
          LazyText: OK (0.24s)
            653  μs ±  59 μs
        map
          Text:     OK (0.24s)
            653  μs ±  57 μs
          LazyText: OK (0.41s)
            653  μs ±  36 μs
        map.map
          Text:     OK (0.41s)
            649  μs ±  30 μs
          LazyText: OK (0.42s)
            656  μs ±  63 μs
        replicate char
          Text:     OK (0.39s)
            22.9 ns ± 2.2 ns
          LazyText: OK (0.35s)
            18.8 ns ± 1.7 ns
        replicate string
          Text:     OK (0.42s)
            27.1 ns ± 2.1 ns
          LazyText: OK (12.71s)
            22.7 ns ± 212 ps
        take
          Text:     OK (0.29s)
            421  μs ±  27 μs
          LazyText: OK (0.48s)
            376  μs ±  17 μs
        tail
          Text:     OK (0.21s)
            572  μs ±  43 μs
          LazyText: OK (0.21s)
            574  μs ±  44 μs
        toLower
          Text:     OK (2.76s)
            90.6 ms ± 7.8 ms
          LazyText: OK (0.76s)
            102  ms ± 4.8 ms
        toUpper
          Text:     OK (0.36s)
            114  ms ± 3.4 ms
          LazyText: OK (0.44s)
            141  ms ± 9.3 ms
        words
          Text:     OK (0.32s)
            20.0 ms ± 776 μs
          LazyText: OK (2.54s)
            38.1 ms ± 903 μs
        zipWith
          Text:     OK (0.25s)
            688  μs ±  60 μs
          LazyText: OK (0.39s)
            589  μs ±  26 μs
    russian
      length
        cons
          Text:     OK (0.16s)
            1.91 μs ± 175 ns
          LazyText: OK (0.17s)
            2.00 μs ± 176 ns
        decode
          Text:     OK (1.17s)
            4.02 μs ± 240 ns
          LazyText: OK (1.11s)
            3.77 μs ±  54 ns
        drop
          Text:     OK (0.29s)
            1.93 μs ±  83 ns
          LazyText: OK (0.15s)
            1.90 μs ± 175 ns
        filter
          Text:     OK (0.20s)
            22.4 μs ± 1.8 μs
          LazyText: OK (0.25s)
            27.5 μs ± 1.9 μs
        filter.filter
          Text:     OK (0.38s)
            22.0 μs ± 1.3 μs
          LazyText: OK (0.13s)
            27.5 μs ± 2.7 μs
        init
          Text:     OK (0.14s)
            1.87 μs ± 172 ns
          LazyText: OK (0.51s)
            1.77 μs ± 118 ns
        intercalate
          Text:     OK (0.22s)
            2.84 μs ± 177 ns
          LazyText: OK (0.22s)
            2.98 μs ± 234 ns
        intersperse
          Text:     OK (0.26s)
            1.81 μs ±  98 ns
          LazyText: OK (0.27s)
            1.81 μs ± 104 ns
        map
          Text:     OK (0.27s)
            1.80 μs ± 105 ns
          LazyText: OK (0.27s)
            1.81 μs ± 111 ns
        map.map
          Text:     OK (0.27s)
            1.81 μs ± 132 ns
          LazyText: OK (0.16s)
            1.88 μs ± 174 ns
        replicate char
          Text:     OK (0.25s)
            23.0 ns ± 1.9 ns
          LazyText: OK (0.20s)
            18.4 ns ± 1.5 ns
        replicate string
          Text:     OK (0.50s)
            26.4 ns ± 1.1 ns
          LazyText: OK (3.41s)
            24.9 ns ± 2.2 ns
        take
          Text:     OK (0.22s)
            1.40 μs ± 105 ns
          LazyText: OK (0.23s)
            1.43 μs ±  88 ns
        tail
          Text:     OK (0.30s)
            2.05 μs ± 107 ns
          LazyText: OK (0.17s)
            2.02 μs ± 178 ns
        toLower
          Text:     OK (0.16s)
            129  μs ±  12 μs
          LazyText: OK (0.76s)
            180  μs ± 7.0 μs
        toUpper
          Text:     OK (0.20s)
            177  μs ±  14 μs
          LazyText: OK (0.51s)
            229  μs ±  12 μs
        words
          Text:     OK (0.73s)
            41.6 μs ± 1.3 μs
          LazyText: OK (0.20s)
            82.8 μs ± 8.1 μs
        zipWith
          Text:     OK (4.05s)
            1.89 μs ±  53 ns
          LazyText: OK (0.14s)
            1.85 μs ± 171 ns
    japanese
      length
        cons
          Text:     OK (0.16s)
            1.98 μs ± 181 ns
          LazyText: OK (0.16s)
            1.97 μs ± 165 ns
        decode
          Text:     OK (1.17s)
            4.08 μs ±  70 ns
          LazyText: OK (1.19s)
            4.11 μs ±  73 ns
        drop
          Text:     OK (1.06s)
            1.90 μs ± 129 ns
          LazyText: OK (0.26s)
            1.87 μs ± 122 ns
        filter
          Text:     OK (0.21s)
            11.8 μs ± 1.1 μs
          LazyText: OK (0.27s)
            15.6 μs ± 892 ns
        filter.filter
          Text:     OK (0.83s)
            12.2 μs ± 554 ns
          LazyText: OK (0.27s)
            15.9 μs ± 919 ns
        init
          Text:     OK (1.14s)
            2.19 μs ± 183 ns
          LazyText: OK (1.13s)
            2.08 μs ±  49 ns
        intercalate
          Text:     OK (0.33s)
            4.46 μs ± 199 ns
          LazyText: OK (0.20s)
            4.98 μs ± 377 ns
        intersperse
          Text:     OK (0.30s)
            2.01 μs ± 107 ns
          LazyText: OK (0.62s)
            2.29 μs ± 180 ns
        map
          Text:     OK (0.61s)
            2.12 μs ± 107 ns
          LazyText: OK (0.64s)
            2.28 μs ± 110 ns
        map.map
          Text:     OK (0.18s)
            2.08 μs ± 182 ns
          LazyText: OK (0.31s)
            2.08 μs ± 198 ns
        replicate char
          Text:     OK (0.46s)
            24.4 ns ± 890 ps
          LazyText: OK (0.72s)
            19.7 ns ± 1.0 ns
        replicate string
          Text:     OK (0.29s)
            28.2 ns ± 1.4 ns
          LazyText: OK (0.25s)
            23.0 ns ± 2.1 ns
        take
          Text:     OK (0.23s)
            1.44 μs ±  85 ns
          LazyText: OK (0.21s)
            1.44 μs ± 139 ns
        tail
          Text:     OK (0.30s)
            2.11 μs ±  90 ns
          LazyText: OK (0.31s)
            2.08 μs ± 182 ns
        toLower
          Text:     OK (0.37s)
            83.3 μs ± 4.4 μs
          LazyText: OK (0.27s)
            116  μs ± 6.1 μs
        toUpper
          Text:     OK (0.70s)
            82.3 μs ± 5.6 μs
          LazyText: OK (0.46s)
            110  μs ± 4.6 μs
        words
          Text:     OK (0.15s)
            59.1 μs ± 5.3 μs
          LazyText: OK (0.20s)
            42.5 μs ± 4.2 μs
        zipWith
          Text:     OK (0.29s)
            1.98 μs ± 110 ns
          LazyText: OK (0.57s)
            2.09 μs ± 204 ns

All 216 tests passed (149.07s)
Benchmark text-benchmarks: FINISH
```

After:

```
Running 1 benchmarks...
Benchmark text-benchmarks: RUNNING...
All
  Pure
    tiny
      length
        cons
          Text:     OK (0.23s)
            23.3 ns ± 1.4 ns
          LazyText: OK (0.23s)
            24.6 ns ± 1.6 ns
        decode
          Text:     OK (0.20s)
            42.0 ns ± 3.4 ns
          LazyText: OK (0.28s)
            126  ns ± 8.0 ns
        drop
          Text:     OK (0.15s)
            32.1 ns ± 2.6 ns
          LazyText: OK (0.17s)
            32.8 ns ± 2.6 ns
        filter
          Text:     OK (0.31s)
            16.2 ns ± 956 ps
          LazyText: OK (0.27s)
            27.9 ns ± 1.4 ns
        filter.filter
          Text:     OK (0.16s)
            16.3 ns ± 1.5 ns
          LazyText: OK (0.14s)
            28.4 ns ± 2.6 ns
        init
          Text:     OK (0.19s)
            20.2 ns ± 1.7 ns
          LazyText: OK (0.24s)
            24.6 ns ± 1.3 ns
        intercalate
          Text:     OK (0.14s)
            26.3 ns ± 2.6 ns
          LazyText: OK (0.14s)
            28.2 ns ± 2.6 ns
        intersperse
          Text:     OK (0.13s)
            26.9 ns ± 2.6 ns
          LazyText: OK (0.26s)
            26.7 ns ± 1.8 ns
        map
          Text:     OK (0.45s)
            24.6 ns ± 1.9 ns
          LazyText: OK (0.26s)
            27.3 ns ± 1.9 ns
        map.map
          Text:     OK (0.25s)
            24.8 ns ± 1.5 ns
          LazyText: OK (0.26s)
            26.2 ns ± 1.3 ns
        replicate char
          Text:     OK (0.44s)
            24.1 ns ± 1.3 ns
          LazyText: OK (0.66s)
            18.4 ns ± 650 ps
        replicate string
          Text:     OK (0.25s)
            24.7 ns ± 1.4 ns
          LazyText: OK (0.22s)
            20.7 ns ± 1.9 ns
        take
          Text:     OK (0.24s)
            23.1 ns ± 1.6 ns
          LazyText: OK (0.28s)
            28.0 ns ± 1.6 ns
        tail
          Text:     OK (0.25s)
            25.0 ns ± 2.2 ns
          LazyText: OK (0.26s)
            25.9 ns ± 1.5 ns
        toLower
          Text:     OK (0.29s)
            117  ns ± 8.2 ns
          LazyText: OK (0.26s)
            208  ns ±  19 ns
        toUpper
          Text:     OK (0.17s)
            121  ns ±  11 ns
          LazyText: OK (0.29s)
            237  ns ±  17 ns
        words
          Text:     OK (0.21s)
            21.6 ns ± 1.4 ns
          LazyText: OK (0.35s)
            36.3 ns ± 2.1 ns
        zipWith
          Text:     OK (0.27s)
            26.7 ns ± 2.4 ns
          LazyText: OK (0.27s)
            28.5 ns ± 1.6 ns
    ascii-small
      length
        cons
          Text:     OK (0.21s)
            22.9 μs ± 2.2 μs
          LazyText: OK (0.20s)
            22.9 μs ± 1.5 μs
        decode
          Text:     OK (1.01s)
            28.5 μs ± 1.8 μs
          LazyText: OK (1.06s)
            29.2 μs ± 1.8 μs
        drop
          Text:     OK (0.22s)
            23.1 μs ± 1.5 μs
          LazyText: OK (0.20s)
            23.0 μs ± 1.4 μs
        filter
          Text:     OK (0.16s)
            136  μs ±  11 μs
          LazyText: OK (0.17s)
            144  μs ±  11 μs
        filter.filter
          Text:     OK (0.13s)
            137  μs ±  11 μs
          LazyText: OK (0.17s)
            144  μs ±  12 μs
        init
          Text:     OK (0.20s)
            22.7 μs ± 1.3 μs
          LazyText: OK (0.21s)
            23.0 μs ± 1.6 μs
        intercalate
          Text:     OK (0.16s)
            33.0 μs ± 2.8 μs
          LazyText: OK (0.17s)
            34.9 μs ± 2.7 μs
        intersperse
          Text:     OK (0.21s)
            22.7 μs ± 1.4 μs
          LazyText: OK (0.22s)
            23.1 μs ± 1.5 μs
        map
          Text:     OK (0.20s)
            21.5 μs ± 1.4 μs
          LazyText: OK (0.20s)
            21.6 μs ± 1.5 μs
        map.map
          Text:     OK (0.20s)
            21.3 μs ± 1.4 μs
          LazyText: OK (0.21s)
            21.8 μs ± 1.5 μs
        replicate char
          Text:     OK (0.23s)
            21.9 ns ± 1.3 ns
          LazyText: OK (0.19s)
            17.5 ns ± 1.4 ns
        replicate string
          Text:     OK (0.25s)
            24.2 ns ± 1.4 ns
…